requirement.txt
opencv-python==4.9.0.80
torch==2.2.2
torchvision==0.17.2
pillow==10.3.0
numpy==1.26.4
argparse==1.4.0
requests==2.31.0
tqdm==4.66.4
diffusers==0.27.2
transformers==4.38.2
accelerate==0.27.2
safetensors==0.4.2